moved number to object type
