+++ src/main/java/io/github/jelastic/core/elastic/ElasticQueryBuilder.java
@@ -18,7 +18,7 @@ package io.github.jelastic.core.elastic;
 
 import io.github.jelastic.core.models.query.filter.FilterVisitor;
 import io.github.jelastic.core.models.query.filter.general.*;
-import io.github.jelastic.core.models.query.filter.number.*;
+import io.github.jelastic.core.models.query.filter.range.*;
 import io.github.jelastic.core.models.query.filter.predicate.AndFilter;
 import io.github.jelastic.core.models.query.filter.predicate.ORFilter;
 import org.elasticsearch.index.query.BoolQueryBuilder;
