+++ src/main/java/io/github/jelastic/core/models/query/filter/range/GreaterEqualsFilter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.jelastic.core.models.query.filter.number;
+package io.github.jelastic.core.models.query.filter.range;
 
 import io.github.jelastic.core.models.query.filter.FilterType;
 import io.github.jelastic.core.models.query.filter.FilterVisitor;
@@ -27,15 +27,15 @@ import lombok.ToString;
 @Data
 @EqualsAndHashCode(callSuper = true)
 @ToString(callSuper = true)
-public class LesserEqualsFilter extends NumberFilter {
+public class GreaterEqualsFilter extends RangeFilter {
 
-    public LesserEqualsFilter() {
-        super(FilterType.LESS_EQUAL);
+    public GreaterEqualsFilter() {
+        super(FilterType.GREATER_EQUAL);
     }
 
-    public LesserEqualsFilter(String field, Number value, boolean includeLower,
-                              boolean includeUpper) {
-        super(FilterType.LESS_EQUAL, field, value, includeLower, includeUpper);
+    public GreaterEqualsFilter(String field, Object value, boolean includeLower,
+                               boolean includeUpper) {
+        super(FilterType.GREATER_EQUAL, field, value, includeLower, includeUpper);
     }
 
     @Override
